requirements.txt
streamlit==1.28.0
tensorflow>=2.16.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
